refactor(faq-dialog): move into pages folder
