+++ src/app/pages/faq/faq-dialog/faq-dialog.component.ts
@@ -1,15 +1,24 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, Inject, OnInit } from '@angular/core';
+import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+
+export interface DialogData {
+  name: string;
+}
 
 @Component({
   selector: 'app-faq-dialog',
   templateUrl: './faq-dialog.component.html',
-  styleUrls: ['./faq-dialog.component.scss']
+  styleUrls: ['./faq-dialog.component.scss'],
 })
 export class FaqDialogComponent implements OnInit {
+  constructor(
+    public dialogRef: MatDialogRef<FaqDialogComponent>,
+    @Inject(MAT_DIALOG_DATA) public data: DialogData
+  ) {}
 
-  constructor() { }
+  ngOnInit(): void {}
 
-  ngOnInit(): void {
+  onNoClick(): void {
+    this.dialogRef.close();
   }
-
 }
